fix migration

+++ backend/alembic/versions/77bf761aee3e_add_is_default_to_mqtt_topics.py
@@ -7,6 +7,7 @@ Create Date: 2025-11-16 03:29:30.124737
 from typing import Sequence, Union
 from alembic import op
 import sqlalchemy as sa
+from sqlalchemy import inspect
 
 # revision identifiers, used by Alembic.
 revision: str = '77bf761aee3e'
@@ -17,9 +18,15 @@ depends_on: Union[str, Sequence[str], None] = None
 
 def upgrade() -> None:
     """Upgrade schema."""
-    op.add_column('mqtt_topics', sa.Column('is_default', sa.Boolean(), nullable=True, server_default="false"))
+    conn = op.get_bind()
+    inspector = inspect(conn)
+    if 'is_default' not in [c['name'] for c in inspector.get_columns('mqtt_topics')]:
+        op.add_column('mqtt_topics', sa.Column('is_default', sa.Boolean(), nullable=True, server_default="false"))
 
 
 def downgrade() -> None:
     """Downgrade schema."""
-    op.drop_column('mqtt_topics', 'is_default')
+    conn = op.get_bind()
+    inspector = inspect(conn)
+    if 'is_default' in [c['name'] for c in inspector.get_columns('mqtt_topics')]:
+        op.drop_column('mqtt_topics', 'is_default')
